+++ src/App.tsx
@@ -9,6 +9,7 @@ import { loadProjects, initProjectListeners } from './store/projects'
 import { initSessionListeners, syncSessionStatuses } from './store/sessions'
 import { initTerminalListeners } from './store/terminals'
 import { initGitListeners, initWindowFocusRefresh } from './store/git'
+import { initSetupListeners } from './store/setup'
 import { loadClaudeSkills } from './store/commands'
 import * as ipc from './lib/ipc'
 import { addToast } from './store/ui'
@@ -43,6 +44,7 @@ const App: Component = () => {
     await initTerminalListeners()
     await initGitListeners()
     await initProjectListeners()
+    await initSetupListeners()
     initWindowFocusRefresh()
     await loadProjects()
     await syncSessionStatuses()